Basic Tic Tac Toe
This is a text based tic tac toe game that is very easy to beat. The
user plays against the computer and the computer simply selects random
cells during its turn. The program as a whole is modularized so that if
a GUI wanted to be used, very little code would need to change.